Set up unified build processes: browser, nodejs
I want to support the 2x2 matrix of:

    Running code...
        ( as executable script | as functions imported to REPL )
        ( in TypeScript via ts-node | in JavaScript via node )

The `yarn test` command exercises all 4 cases in the above 2x2 matrix.

---

About `baseUrl`:

To support writing imports like

    import { srcPath } from 'src/nodejs-env';

in TypeScript source, you would at least need to set

    "baseUrl": "./"

in `tsconfig.json` --- however, that makes the build output in `ts-out/`
not executable via `node`, because node still tries to resolve modules
via literally `src/` when you execute:

    node ts-out/nodejs-main.js

The solution is to always use lots of double-dots, and not use `baseUrl`
at all; as such, this commit removes the `baseUrl` option from my
`tsconfig.json`.

If double-dots are undesirable, then the `tsconfig-paths` npm package is
needed, with an extra no-op value for the `paths` option. A file like
this would be needed:

    // register-tsconfig-paths.js
    const tsConfig = require("./tsconfig.json");
    const tsConfigPaths = require("tsconfig-paths");
    tsConfigPaths.register({
      baseUrl: tsConfig.compilerOptions.baseUrl,
      paths: {
          '*': ['*']
      } // Or write this in tsConfig.compilerOptions.paths
    });

and to run the nodejs interpreter on TypeScript compilation output, you
would also need to loop in `tsconfig-paths` as a runtime dependency:

    node -r ./register-tsconfig-paths.js ts-out/nodejs-main.js

---

About module format:

In this commit, I've switched all module formats to a very conservative
CommonJS module format. Rollup is somehow smart/weird enough to be able
to produce an ES2015-module bundle even when given CommonJS input.

The reason to use CommonJS as the output module format for TypeScript
compilation is so that I can conveniently run NodeJS tools on `ts-out/`
files without any configuration. There are lots of weird options like
`type: module`:

    // package.json
    {
        "name": "twitanalysis-frontend",
        "type": "module",
        ...
    }

However, setting this option has widespread implications, e.g. somehow
making `ts-node` unable to interpret TypeScript library module source
files no matter how you write the module source-code.

---

About `require.main === module`:

There is definitely this weird nodejs-ism in `src/nodejs-main.ts`
TypeScript source, already written in ES2015 module format; but it's not
so important, because any module which requires such semantics are
intended to be interpreted by `node`, which will have the correct
globals as expected. It seems to work naturally, so long as the
TypeScript compiler doesn't have an issue with it --- and so far, things
appear to be working as-is.