Set up Babel for transpiling some jsx
The `@rollup/plugin-babel` npm package README didn't explicitly mention
the `presets:` option for some reason, but did use it in their sample
code. Trying out a common-sense config seems to make things work.

- Rollup: https://www.npmjs.com/package/@rollup/plugin-babel
- Babel: https://babeljs.io/docs/en/usage

Tried out an ad-hoc command like this, which seems to produce a js file
which I'm not sure works or not:

    yarn babel \
        --presets=@babel/preset-env,@babel/preset-react \
        --out-dir babel-out/ \
        static/react-playground.js  # This file was before this commit